Add MapRenderAPI with core functionality.
The API class can receive a list of addresses and render/save the resulting route image.
Waypoint markers are being implemented.
MapRenderManager is still in the works as there is no clean way of representing a day's worth of classes yet.
Several ArrayList declarations have been changed to List as good practice.